+++ projects/packages/premium-analytics/packages/data/src/processing/stats/__tests__/comments.test.ts
@@ -16,8 +16,8 @@ describe( 'Stats comments normalizer', () => {
 		expect( result.data ).toEqual( [
 			expect.objectContaining( {
 				time_interval: '2026-06-16',
-				date_start: '2026-06-16T00:00:00+00:00',
-				date_end: '2026-06-16T23:59:59+00:00',
+				date_start: '2026-06-16T00:00:00',
+				date_end: '2026-06-16T23:59:59',
 				items: [
 					{
 						label: 'authors',